Moved Disk class definition to its own file
