+++ archey/archey.py
@@ -18,6 +18,7 @@ from hostname import Hostname
 from configuration import Configuration
 from kernel import Kernel
 from uptime import Uptime
+from disk import Disk
 from constants import (
     COLOR_DICT,
     DE_DICT,
@@ -336,30 +337,6 @@ class RAM:
             int(used),
             COLOR_DICT['clear'],
             int(total)
-        )
-
-
-class Disk:
-    def __init__(self):
-        total = re.sub(
-            ',', '.',
-            check_output(
-                [
-                    'df', '-Tlh', '-B', 'GB', '--total',
-                    '-t', 'ext4', '-t', 'ext3', '-t', 'ext2',
-                    '-t', 'reiserfs', '-t', 'jfs', '-t', 'zfs',
-                    '-t', 'ntfs', '-t', 'fat32', '-t', 'btrfs',
-                    '-t', 'fuseblk', '-t', 'xfs', '-t', 'simfs',
-                    '-t', 'tmpfs', '-t', 'lxfs'
-                ], universal_newlines=True
-            ).splitlines()[-1]
-        ).split()
-
-        self.value = '{0}{1}{2} / {3}'.format(
-            COLOR_DICT['sensors'][int(float(total[5][:-1]) // 33.34)],
-            re.sub('GB', ' GB', total[3]),
-            COLOR_DICT['clear'],
-            re.sub('GB', ' GB', total[2])
         )
